Firmware: vendor the modem library with fixes; bug fixes; 32-bit counter epoch
docs/V02_DESIGN.md sections 1-3 (firmware half).

Modem library: dptechnics/walter-modem v1.5.0 moves from the component
manager into firmware/components/ so it can be patched (licence permits
modified source, Walter boards only; LICENSE and headers kept). Patches,
each marked PAGER PATCH and listed in PATCHES.md:
- no _finishModemCMD(NULL): a stray NO CARRIER used to reboot the pager
- RX buffer writes and the MQTT/socket receive copies are bounded by the
  1540-byte buffer (the raw UART ingest path could write past it)
- TLS profile ids up to 6, the modem's real limit

Bugs:
- acks: ack_state only advances when the ack was actually queued, and a
  render marks every unshown message, so a burst no longer loses acks
- ident_store() updates the in-memory identity (epoch was one behind)
- connect watchdog: 60 s with no result -> disconnect and back off;
  three in a row -> modem recovery
- placeholder CA (ISRG Root X2) written to slot 12 when nothing is
  pinned, so an empty slot can never trigger the plaintext fallback
- malformed /down drops are logged at INFO with length and first byte
- debug build only: the console REPL also runs on a provisioned pager

Counter: n = (epoch << 20) | lo with a 32-bit epoch (was 12), uint64 on
the up path, NVS migration n_epoch (u16) -> n_epoch32. The down path
parses a wide n without rejecting it. Host vectors for n >= 2^32.

Host tests pass; normal and debug idf.py builds are clean. Not run on
hardware.

+++ firmware/main/auth.c
@@ -291,10 +291,13 @@ bool auth_verify(const char *topic, uint8_t *buf, size_t *len)
  * Replay counters.
  * --------------------------------------------------------------------- */
 
-uint32_t auth_next_up_n(auth_rtc_t *rtc, uint16_t epoch, bool *wrapped)
+uint64_t auth_next_up_n(auth_rtc_t *rtc, uint32_t epoch, bool *wrapped)
 {
     uint32_t lo = rtc->up_lo & AUTH_UP_LO_MASK;
-    uint32_t n = (((uint32_t) (epoch & AUTH_UP_EPOCH_MASK)) << AUTH_UP_LO_BITS) | lo;
+    /* v0.2 (docs/V02_DESIGN.md §3): epoch is the full 32 bits now, so the
+     * shift must happen in 64-bit arithmetic (a plain uint32_t << 20 would
+     * silently drop the top 20 bits of a large epoch). */
+    uint64_t n = (((uint64_t) (epoch & AUTH_UP_EPOCH_MASK)) << AUTH_UP_LO_BITS) | lo;
 
     uint32_t next_lo = (lo + 1u) & AUTH_UP_LO_MASK;
     bool did_wrap = (next_lo == 0u); /* wrapped past AUTH_UP_LO_MASK back to 0 */
